+++ lib/models/address/grx_address_form_strings.dart
@@ -1,0 +1,40 @@
+/// Localized strings required by [GrxAddressForm].
+abstract class GrxAddressFormStrings {
+  const GrxAddressFormStrings();
+
+  String get addressSearchError;
+  String get addressPostalCodeInvalidBrCep;
+  String get cepNotFound;
+  String get invalidField;
+  String get addressLine1Label;
+  String get addressLine2Label;
+  String get dependentLocality;
+  String get locality;
+  String get administrativeArea;
+  String get postalCode;
+  String get country;
+  String get countryBR;
+  String get countryUY;
+  String get countryAR;
+  String get countryPY;
+  String get countryMZ;
+  String get countryUS;
+  String get countryPT;
+  String get countryES;
+  String get addressNumberRequired;
+  String get enterAdministrativeArea;
+  String get addressLine1Placeholder;
+  String get addressLine2Placeholder;
+  String get addressDependentLocalityPlaceholder;
+  String get addressLocalityPlaceholder;
+  String get addressAdministrativeAreaPlaceholder;
+  String get addressPostalCodePlaceholder;
+  String get addAddressNumberHelper;
+  String get searchCep;
+  String get enterYourCountry;
+  String get searchAddress;
+  String get enterAddressSearch;
+  String get noAddressResults;
+  String get addressSearchHelperText;
+  String get removeAddress;
+}