+++ lib/widgets/bottom_sheet/grx_bottom_sheet_form_field_body.widget.dart
@@ -1,6 +1,9 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 
 import '../../themes/colors/grx_colors.dart';
+import '../../themes/spacing/grx_spacing.dart';
 import '../../utils/grx_text_sanitizer.util.dart';
 import '../buttons/grx_primary_button.widget.dart';
 import '../buttons/grx_secondary_button.widget.dart';
@@ -27,16 +30,11 @@ class GrxBottomSheetFormFieldBody<T> extends StatefulWidget {
     this.emptyListText,
     this.confirmButtonLabel,
     this.cancelButtonLabel,
-  }) : assert(
-         (!multiSelect && onChangeState != null) ||
-             (multiSelect && valueKey != null),
-       ),
+    this.includeListBottomSafeAreaPadding = true,
+  }) : assert((!multiSelect && onChangeState != null) || (multiSelect && valueKey != null)),
        assert(
          !searchable ||
-             (searchable &&
-                 quickSearchFieldController != null &&
-                 onFilterSetState != null &&
-                 displayText != null),
+             (searchable && quickSearchFieldController != null && onFilterSetState != null && displayText != null),
        );
 
   final ScrollController? controller;
@@ -44,13 +42,7 @@ class GrxBottomSheetFormFieldBody<T> extends StatefulWidget {
   final String Function(T data)? displayText;
   final TextEditingController? quickSearchFieldController;
   final void Function(T?)? onSelectItem;
-  final Widget Function(
-    BuildContext context,
-    int index,
-    T value,
-    void Function()? onChanged,
-    bool isSelected,
-  )
+  final Widget Function(BuildContext context, int index, T value, void Function()? onChanged, bool isSelected)
   itemBuilder;
   final Iterable<T> items;
   final void Function(T)? onChangeState;
@@ -63,15 +55,19 @@ class GrxBottomSheetFormFieldBody<T> extends StatefulWidget {
   final String? emptyListText;
   final String? confirmButtonLabel;
   final String? cancelButtonLabel;
+  final bool includeListBottomSafeAreaPadding;
 
   @override
   State<StatefulWidget> createState() => _GrxBottomSheetFormFieldBodyState<T>();
 }
 
-class _GrxBottomSheetFormFieldBodyState<T>
-    extends State<GrxBottomSheetFormFieldBody<T>> {
+class _GrxBottomSheetFormFieldBodyState<T> extends State<GrxBottomSheetFormFieldBody<T>> {
+  static const _kSearchFieldHeight = 48.0;
+  static const _kSearchDebounceDuration = Duration(milliseconds: 500);
+
   final _list = <T>[];
   final _selectedValues = <T>[];
+  Timer? _debounceTimer;
 
   @override
   void initState() {
@@ -83,6 +79,12 @@ class _GrxBottomSheetFormFieldBodyState<T>
     if (widget.initialSelectedValues != null) {
       _selectedValues.addAll(widget.initialSelectedValues!);
     }
+  }
+
+  @override
+  void dispose() {
+    _debounceTimer?.cancel();
+    super.dispose();
   }
 
   void _onItemCheckedChange(T itemValue) {
@@ -97,17 +99,17 @@ class _GrxBottomSheetFormFieldBodyState<T>
     });
   }
 
+  void _onSearchChanged(String value) {
+    _debounceTimer?.cancel();
+    _debounceTimer = Timer(_kSearchDebounceDuration, () => _filterData(value));
+  }
+
   void _filterData(String val) {
     void filter() {
       _list.clear();
       _list.addAll(
         widget.items.where(
-          (x) =>
-              val.isEmpty ||
-              GrxTextSanitizer.matchesSearch(
-                val,
-                widget.displayText!(x).toString(),
-              ),
+          (x) => val.isEmpty || GrxTextSanitizer.matchesSearch(val, widget.displayText!(x).toString()),
         ),
       );
     }
@@ -119,127 +121,115 @@ class _GrxBottomSheetFormFieldBodyState<T>
         : filter();
   }
 
+  Widget _buildFloatingSearchSliver() {
+    return SliverAppBar(
+      primary: false,
+      floating: true,
+      snap: true,
+      pinned: false,
+      toolbarHeight: 0,
+      automaticallyImplyLeading: false,
+      backgroundColor: GrxColors.background,
+      surfaceTintColor: Colors.transparent,
+      scrolledUnderElevation: 0,
+      elevation: 0,
+      bottom: PreferredSize(
+        preferredSize: const Size.fromHeight(GrxSpacing.xs + _kSearchFieldHeight + GrxSpacing.s),
+        child: Padding(
+          padding: const EdgeInsets.fromLTRB(GrxSpacing.s, GrxSpacing.xs, GrxSpacing.s, GrxSpacing.s),
+          child: GrxSearchField(
+            searchFieldController: widget.quickSearchFieldController!,
+            onChanged: _onSearchChanged,
+            hintText: widget.searchHintText ?? 'Search',
+          ),
+        ),
+      ),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     final mediaQuery = MediaQuery.of(context);
+    final listBottomPadding =
+        widget.multiSelect
+            ? 0.0
+            : widget.includeListBottomSafeAreaPadding
+            ? mediaQuery.viewInsets.bottom + mediaQuery.padding.bottom + GrxSpacing.s
+            : 0.0;
+
+    final listContent = ColoredBox(
+      color: GrxColors.background,
+      child: CustomScrollView(
+        controller: widget.controller,
+        shrinkWrap: widget.shrinkWrap,
+        slivers: [
+          if (widget.searchable) ...[_buildFloatingSearchSliver()],
+          _list.isEmpty
+              ? SliverToBoxAdapter(
+                child: SafeArea(
+                  top: false,
+                  child: Center(
+                    child: Padding(
+                      padding: const EdgeInsets.all(16.0),
+                      child: GrxLabelLargeText(widget.emptyListText ?? 'No results found'),
+                    ),
+                  ),
+                ),
+              )
+              : SliverPadding(
+                padding: EdgeInsets.only(
+                  left: GrxSpacing.s,
+                  top: GrxSpacing.s,
+                  right: GrxSpacing.s,
+                  bottom: listBottomPadding,
+                ),
+                sliver: SliverList(
+                  delegate: SliverChildBuilderDelegate(childCount: _list.length, (context, index) {
+                    final item = _list[index];
+
+                    if (widget.multiSelect) {
+                      return widget.itemBuilder(
+                        context,
+                        index,
+                        item,
+                        () => _onItemCheckedChange(item),
+                        _selectedValues.any((element) => widget.valueKey!(element) == widget.valueKey!(item)),
+                      );
+                    }
+
+                    return InkWell(
+                      onTap: () {
+                        Navigator.pop(context, true);
+
+                        if (widget.onSelectItem != null) {
+                          widget.onSelectItem!(item);
+                        }
+                        widget.onChangeState!(item);
+                      },
+                      child: widget.itemBuilder(context, index, item, null, false),
+                    );
+                  }),
+                ),
+              ),
+        ],
+      ),
+    );
 
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: [
-        Flexible(
-          child: ColoredBox(
-            color: GrxColors.background,
-            child: CustomScrollView(
-              controller: widget.controller,
-              shrinkWrap: widget.shrinkWrap,
-              slivers: [
-                if (widget.searchable)
-                  SliverAppBar(
-                    titleSpacing: 0.0,
-                    toolbarHeight: 65.0,
-                    backgroundColor: Colors.transparent,
-                    automaticallyImplyLeading: false,
-                    title: Container(
-                      color: GrxColors.background,
-                      padding: const EdgeInsets.symmetric(
-                        vertical: 10,
-                        horizontal: 8,
-                      ),
-                      child: GrxSearchField(
-                        searchFieldController:
-                            widget.quickSearchFieldController!,
-                        onChanged: _filterData,
-                        hintText: widget.searchHintText ?? 'Search',
-                      ),
-                    ),
-                  ),
-                _list.isEmpty
-                    ? SliverToBoxAdapter(
-                      child: SafeArea(
-                        top: false,
-                        child: Center(
-                          child: Padding(
-                            padding: const EdgeInsets.all(16.0),
-                            child: GrxLabelLargeText(
-                              widget.emptyListText ?? 'No results found',
-                            ),
-                          ),
-                        ),
-                      ),
-                    )
-                    : SliverPadding(
-                      padding: EdgeInsets.only(
-                        left: 8,
-                        top: 8,
-                        right: 8,
-                        bottom:
-                            !widget.multiSelect
-                                ? MediaQuery.of(context).viewInsets.bottom +
-                                    MediaQuery.of(context).padding.bottom +
-                                    8
-                                : 0,
-                      ),
-                      sliver: SliverList(
-                        delegate: SliverChildBuilderDelegate(
-                          childCount: _list.length,
-                          (context, index) {
-                            final item = _list.toList()[index];
-
-                            if (widget.multiSelect) {
-                              return widget.itemBuilder(
-                                context,
-                                index,
-                                item,
-                                () => _onItemCheckedChange(item),
-                                _selectedValues.any(
-                                  (element) =>
-                                      widget.valueKey!(element) ==
-                                      widget.valueKey!(item),
-                                ),
-                              );
-                            } else {
-                              return InkWell(
-                                onTap: () {
-                                  Navigator.pop(context, true);
-
-                                  if (widget.onSelectItem != null) {
-                                    widget.onSelectItem!(item);
-                                  }
-                                  widget.onChangeState!(item);
-                                },
-                                child: widget.itemBuilder(
-                                  context,
-                                  index,
-                                  item,
-                                  null,
-                                  false,
-                                ),
-                              );
-                            }
-                          },
-                        ),
-                      ),
-                    ),
-              ],
-            ),
-          ),
-        ),
+        Flexible(child: listContent),
         if (widget.multiSelect)
           Container(
             decoration: BoxDecoration(
               color: GrxColors.neutrals,
-              border: Border(
-                top: BorderSide(color: GrxColors.primary.shade50, width: 1),
-              ),
+              border: Border(top: BorderSide(color: GrxColors.primary.shade50, width: 1)),
             ),
             padding: EdgeInsets.only(
               top: 12.0,
               right: 16.0,
               left: 16.0,
-              bottom:
-                  mediaQuery.viewInsets.bottom +
-                  mediaQuery.padding.bottom +
-                  12.0,
+              bottom: mediaQuery.viewInsets.bottom + mediaQuery.padding.bottom + 12.0,
             ),
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,